Consolidate service descriptor alteration
