+++ src/Components/Aspire.Microsoft.EntityFrameworkCore.SqlServer/AspireSqlServerEFCoreSqlClientExtensions.cs
@@ -108,38 +108,10 @@ public static class AspireSqlServerEFCoreSqlClientExtensions
                 return;
             }
 
-            // Resolving DbContext<TContextService> will resolve DbContextOptions<TContextImplementation>.
-            // We need to replace the DbContextOptions service descriptor to inject more logic. This won't be necessary once
-            // Aspire targets .NET 9 as EF will respect the calls to services.ConfigureDbContext<TContext>(). c.f. https://github.com/dotnet/efcore/pull/32518
-
-            var oldDbContextOptionsDescriptor = builder.Services.FirstOrDefault(sd => sd.ServiceType == typeof(DbContextOptions<TContext>));
-
-            if (oldDbContextOptionsDescriptor is null)
+            builder.PatchServiceDescriptor<TContext>(optionsBuilder =>
             {
-                throw new InvalidOperationException($"DbContext<{typeof(TContext).Name}> was not registered. Ensure you have registered the DbContext in DI before calling {nameof(EnrichSqlServerDbContext)}.");
-            }
-
-            builder.Services.Remove(oldDbContextOptionsDescriptor);
-
-            var dbContextOptionsDescriptor = new ServiceDescriptor(
-                oldDbContextOptionsDescriptor.ServiceType,
-                oldDbContextOptionsDescriptor.ServiceKey,
-                factory: (sp, key) =>
-                {
-                    var dbContextOptions = oldDbContextOptionsDescriptor.ImplementationFactory?.Invoke(sp) as DbContextOptions<TContext>;
-
-                    var optionsBuilder = dbContextOptions != null
-                        ? new DbContextOptionsBuilder<TContext>(dbContextOptions)
-                        : new DbContextOptionsBuilder<TContext>();
-
-                    optionsBuilder.UseSqlServer(options => options.EnableRetryOnFailure());
-
-                    return optionsBuilder.Options;
-                },
-                oldDbContextOptionsDescriptor.Lifetime
-                );
-
-            builder.Services.Add(dbContextOptionsDescriptor);
+                optionsBuilder.UseSqlServer(options => options.EnableRetryOnFailure());
+            });
         }
     }
