+++ src/Components/Aspire.Npgsql.EntityFrameworkCore.PostgreSQL/AspireEFPostgreSqlExtensions.cs
@@ -115,38 +115,7 @@ public static partial class AspireEFPostgreSqlExtensions
                 return;
             }
 
-            // Resolving DbContext<TContextService> will resolve DbContextOptions<TContextImplementation>.
-            // We need to replace the DbContextOptions service descriptor to inject more logic. This won't be necessary once
-            // Aspire targets .NET 9 as EF will respect the calls to services.ConfigureDbContext<TContext>(). c.f. https://github.com/dotnet/efcore/pull/32518
-
-            var oldDbContextOptionsDescriptor = builder.Services.FirstOrDefault(sd => sd.ServiceType == typeof(DbContextOptions<TContext>));
-
-            if (oldDbContextOptionsDescriptor is null)
-            {
-                throw new InvalidOperationException($"DbContext<{typeof(TContext).Name}> was not registered. Ensure you have registered the DbContext in DI before calling {nameof(EnrichNpgsqlDbContext)}.");
-            }
-
-            builder.Services.Remove(oldDbContextOptionsDescriptor);
-
-            var dbContextOptionsDescriptor = new ServiceDescriptor(
-                oldDbContextOptionsDescriptor.ServiceType,
-                oldDbContextOptionsDescriptor.ServiceKey,
-                factory: (sp, key) =>
-                {
-                    var dbContextOptions = oldDbContextOptionsDescriptor.ImplementationFactory?.Invoke(sp) as DbContextOptions<TContext>;
-
-                    var optionsBuilder = dbContextOptions != null
-                        ? new DbContextOptionsBuilder<TContext>(dbContextOptions)
-                        : new DbContextOptionsBuilder<TContext>();
-
-                    optionsBuilder.UseNpgsql(options => options.EnableRetryOnFailure());
-
-                    return optionsBuilder.Options;
-                },
-                oldDbContextOptionsDescriptor.Lifetime
-                );
-
-            builder.Services.Add(dbContextOptionsDescriptor);
+            builder.PatchServiceDescriptor<TContext>(optionsBuilder => optionsBuilder.UseNpgsql(options => options.EnableRetryOnFailure()));
         }
     }
