+++ src/Components/Aspire.Pomelo.EntityFrameworkCore.MySql/AspireEFMySqlExtensions.cs
@@ -137,45 +137,18 @@ public static partial class AspireEFMySqlExtensions
                 return;
             }
 
-            // Resolving DbContext<TContextService> will resolve DbContextOptions<TContextImplementation>.
-            // We need to replace the DbContextOptions service descriptor to inject more logic. This won't be necessary once
-            // Aspire targets .NET 9 as EF will respect the calls to services.ConfigureDbContext<TContext>(). c.f. https://github.com/dotnet/efcore/pull/32518
-
-            var oldDbContextOptionsDescriptor = builder.Services.FirstOrDefault(sd => sd.ServiceType == typeof(DbContextOptions<TContext>));
-
-            if (oldDbContextOptionsDescriptor is null)
+            builder.PatchServiceDescriptor<TContext>(optionsBuilder =>
             {
-                throw new InvalidOperationException($"DbContext<{typeof(TContext).Name}> was not registered. Ensure you have registered the DbContext in DI before calling {nameof(EnrichMySqlDbContext)}.");
-            }
-
-            builder.Services.Remove(oldDbContextOptionsDescriptor);
-
-            var dbContextOptionsDescriptor = new ServiceDescriptor(
-                oldDbContextOptionsDescriptor.ServiceType,
-                oldDbContextOptionsDescriptor.ServiceKey,
-                factory: (sp, key) =>
-                {
-                    var dbContextOptions = oldDbContextOptionsDescriptor.ImplementationFactory?.Invoke(sp) as DbContextOptions<TContext>;
-
 #pragma warning disable EF1001 // Internal EF Core API usage.
-                    if (dbContextOptions is null ||
-                        dbContextOptions.FindExtension<MySqlOptionsExtension>() is not MySqlOptionsExtension extension ||
-                        extension.ServerVersion is not ServerVersion serverVersion)
-                    {
-                        throw new InvalidOperationException($"A DbContextOptions<{typeof(TContext).Name}> was not found. Please ensure 'ServerVersion' was configured.");
-                    }
+                if (optionsBuilder.Options.FindExtension<MySqlOptionsExtension>() is not MySqlOptionsExtension extension
+                   || extension.ServerVersion is not ServerVersion serverVersion)
+                {
+                    throw new InvalidOperationException($"A DbContextOptions<{typeof(TContext).Name}> was not found. Please ensure 'ServerVersion' was configured.");
+                }
 #pragma warning restore EF1001 // Internal EF Core API usage.
 
-                    var optionsBuilder = new DbContextOptionsBuilder<TContext>(dbContextOptions);
-
-                    optionsBuilder.UseMySql(serverVersion, options => options.EnableRetryOnFailure());
-
-                    return optionsBuilder.Options;
-                },
-                oldDbContextOptionsDescriptor.Lifetime
-                );
-
-            builder.Services.Add(dbContextOptionsDescriptor);
+                optionsBuilder.UseMySql(serverVersion, options => options.EnableRetryOnFailure());
+            });
         }
     }
